+++ waspc/cli/src/Wasp/Cli/Command/BuildStart/Config.hs
@@ -22,6 +22,7 @@ import Wasp.Cli.Util.Apps (getDevUrlMakers, getWaspEnvVars)
 import Wasp.Cli.Util.Parser (getParserHelpMessage)
 import Wasp.Cli.Util.PathArgument (FilePathArgument)
 import qualified Wasp.Cli.Util.PathArgument as PathArgument
+import qualified Wasp.Cli.Util.PortArgument as PortArgument
 import Wasp.Env (EnvVar, nubEnvVars, overrideEnvVars, parseDotEnvFile)
 import Wasp.Generator.Common (GeneratedAppDir)
 import Wasp.Project.Apps (Apps (..))
@@ -39,7 +40,7 @@ data BuildStartConfig = BuildStartConfig
 
 makeBuildStartConfig :: AppSpec -> BuildStartArgs -> SP.Path' SP.Abs (SP.Dir WaspProjectDir) -> Command BuildStartConfig
 makeBuildStartConfig appSpec args projectDir' = do
-  let ports = args.ports
+  ports <- PortArgument.resolveAppPorts args.ports
 
   userEnvVars <- liftIO $ traverse combineEnvVarsWithEnvFiles args.envInputs
   when (all null userEnvVars) $ throwError noEnvVarsSpecifiedMsg